Handle rename collisions
There may be cases so that a file A in Windows is called B in Linux, and B in Windows is called A in Linux. Neither file can be renamed before the other without a temporary filename.

As a workaround, all files are now moved under a temporary folder and then moved to the final position only after all Windows-only files and directories have been deleted. This way, Linux names are introduced when all Windows names are gone and there can be no clashes.

This is done systematically, even if there are no clashes given the two pathnames sets. Also, the temporary directory name is created in advance by gogdiff and stored as a fixed value in the delta script, as it is assumed that it will also be unique when extracting (since we assume the same layout for the target Windows directory).